Handle special case when comparing LPCM mime types
Based on a patch from Yakup Akbay <yakbay@ubicom.com>. The problem with
his patch was that it was reporting LPCM with different parameter values
as compatible. For example, "audio/L16;rate=44100;channels=2" and
"audio/L16;rate=4100;channels=1". OTOH, this patch assumes the order of
these parameters to be the same in mime-type strings being compared. A
more complete solution would be to compare each mimetype parameter
separately.